requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pymongo>=4.6.0